--- src/app/page.tsx
@@ -1,14 +1,0 @@
-import { AboutSection, HeroSection } from '@/components/ui';
-
-const Home = () => {
-  return (
-    <div className="flex flex-col items-center">
-      <HeroSection />
-      <AboutSection />
-      {/* <ExperienceSection /> */}
-      {/* <ContactSection /> */}
-    </div>
-  );
-};
-
-export default Home;
